Redefine the exception's "type"
new: Re-add support for custom exceptions, with unique message. The id will be 'framework#0' (unknown exception) in this cases, and there is no more `Exception\Helper::EXCEPTION_INVALID_ID`
update!: Remove the `Exception->type`, but keep the postfix in the exception id. The postfix now only affect the log level of the exception, and not included in the `Exception->id`