Refactor and modifications in Logging
!refactor: Extract file logging from `Logger` and move it to `Logger\File`, while replacing `LoggerVoid` with the newly cleaned `Logger` class which is now a stores only to the memory. From now it's easier to extend the Logger with different storage mechanisms
new: The `Logger` class uses a memory buffer for the log entries, to be able to flush it in batch which can improve the performance. For example the new `Logger\File` will flush into a file only on destruct or manually with the `->flush()` method. The buffered entries can be dropped with the `->clear()` method to prevent storing it permanently
new: Add support for callable context in the `Logger` class, to be able to delay the context computations to happen only if the log entry is stored in the log buffer. This can improve performance in some cases
update: The `Logger::wrap()` has been renamed to `::dump()` (and became static to be able to use elsewhere much easier) and improved with `__debugInfo` usage if the dumped object supports it